Fix: edit the functions names in finalizer.go in finalizer packages , compiled the code.
